==> docs/04-createData-vacancy.Rmd <==
---
title: "Create Data - Vacancy"
author: "Christopher Prener, Ph.D."
date: '(`r format(Sys.time(), "%B %d, %Y")`)'
output: 
  github_document: default
  html_notebook: default 
---

## Introduction
This notebook creates the grid-level vacancy estimates for further analyses. These spatial data will also be used as a base to interpolate demographic data into.

## Dependencies
This notebook requires the following packages:

```{r load-packages}
# tidyverse packages
library(dplyr)
library(readr)

# spatial packages
library(sf)

# other packages
library(here)
```

## Open Grid Data
The grid data have a large number of variables from a prior study. We'll start by cleaning them up on import:

```{r open-grids}
vacancy <- st_read(here("data", "raw", "vacancy"), stringsAsFactors = FALSE) %>%
  select(OBJECTID_1, sq_m, v_sq_m, p_vacant) %>%
  rename(
    grid_id = OBJECTID_1,
    area_m2 = sq_m,
    vacant_m2 = v_sq_m,
    prop_vacant = p_vacant
  )
```

Next, we'll create a copy that only contains the `grid_id` column. Since shapefile variable names get truncated, we'll store the main analytical data set as a `.csv` file separate from the geometric data:

```{r}
# make copy
grids <- vacancy %>%
  select(grid_id)

# re-project
grids <- st_transform(grids, crs = 26915)

# remove geometry
st_geometry(vacancy) <- NULL
```


## Store Copy
Since the goal of the `data/clean/` folder is to keep a record of all the individual data sets being used in the project, I'll write a copy of the `grids` object into that subdirectory:

```{r write-grids}
st_write(obj = grids, dsn = here("data", "raw", "grids", "grids.shp"), delete_dsn = TRUE)
```

We'll also write the vacancy data as a `.csv` file:

```{r write-vacancy}
write_csv(vacancy, path = here("data", "clean", "grids_vacancy.csv"))
```
