==> docs/09-analysis-longitudinal.Rmd <==
---
title: "Analysis - Longitudinal Data"
author: "Christopher Prener, Ph.D."
date: '(`r format(Sys.time(), "%B %d, %Y")`)'
output: github_document
---

## Introduction
This notebook performs the longitudinal analyses for the arson data.

## Dependencies
The following packages are needed:

```{r load-dependencies}
# time series
library(forecast)
library(ggfortify)

# tidyverse
library(ggplot2)
library(readr)

# other packages
library(heplots)
library(here)
library(scales)
```

## Load Data
The data for these analyses are:

```{r load-data}
long <- read_csv(here("data", "analysis", "longitudinal_analysis.csv"))
```


## Plots

```{r plot-arson}
ggplot(long, mapping = aes(x = year, y = arson_rate)) +
  geom_line() +
  scale_x_continuous(breaks = c(1980, 1985, 1990, 1995, 2000, 2005, 2010, 2015, 2018)) +
  scale_y_continuous(limits = c(0, 250)) +
  labs(
    title = "Arson Rate in St. Louis, 1980-2018",
    x = "Year",
    y = "Rate per 100,000 Residents"
  ) -> plot

ggsave(filename = here("results", "rate.png"), width = 7, height = 4, units = "in", dpi = 500)

plot
```


```{r}
ggplot(long) +
  geom_line(mapping = aes(x = year, y = pop)) +
  geom_line(mapping = aes(x = year, y = manu_jobs)) + 
  scale_x_continuous(breaks = c(1980, 1985, 1990, 1995, 2000, 2005, 2010, 2015, 2018)) +
  scale_y_continuous(labels = comma) +
  labs(
    title = "Demographic Declines, 1980-2018",
    x = "Year",
    y = "Count"
  )
```

## Convet to TS Object

```{r}
long <- ts(long, start = 1980, end = 2018, frequency = 1)
```

```{r}
# autoplot(long[,c("arson_rate", "violent_rate")])
```


```{r}
m1 <- tslm(arson_rate ~ violent_rate + larceny_rate, data = long)

summary(m1)
```

```{r}
m2 <- tslm(arson_rate ~ pop_delta_pct + manu_delta_pct, data = long)

summary(m2)
```

```{r}
m2 <- tslm(arson_rate ~ pop_delta_pct + manu_delta_pct, data = long)

summary(m2)
```

```{r}
m3 <- tslm(arson_rate ~ violent_rate + larceny_rate + pop_delta_pct + manu_delta_pct, data = long)

summary(m3)
```

```{r}
etasq(m3, partial = FALSE)
```
