==> docs/10-analysis-spatial.Rmd <==
---
title: "Analysis - Spatial Data"
author: "Christopher Prener, Ph.D."
date: '(`r format(Sys.time(), "%B %d, %Y")`)'
output: 
  github_document: default
  html_notebook: default 
---

## Introduction
This notebook fits some experimental zero-inflated models using eigenvector filtering to reduce the impact of spatial auto-correlation on the results.

## Dependencies
This notebook requires the following packages:

```{r load-packages}
# tidyverse packages
library(dplyr)

# spatial packages
library(sf)
library(sp)
library(spdep)

# modeling packages
library(AER)
library(MASS)
library(pscl)

# other packages
library(here)
```

## Load Data
This notebook requires the final data set:

```{r load-data}
grid <- st_read(here("data", "analysis", "spatial_analysis.geojson"), crs = 26915)
```

## Caluclate Pop Change
We'll calculate the proportion of change from 1950 through 2017 and 1970 through 2017:

```{r pop-change}
grid %>%
  mutate(delta50 = (total_pop-pop50)/pop50) %>%
  mutate(delta70 = total_pop-pop70/pop70) -> grid
```

There are two grids with missing data:

```{r}
grid <- filter(grid, is.na(pop70) == FALSE)
```


## Weights Matrix
We also need a spatial weights matrix for our models. We'll use rook since we have gridded data:

```{r weights-matrix}
weights <- poly2nb(grid, queen=FALSE)
weights <- nb2listw(weights, style="C", zero.policy=TRUE)
```

## Moran's I
Next, we'll use our gridded data and our spatial weights matrix to calculate Moran's $I$, which is a measure of spatial autocorrelation:

```{r evaluate-i}
moran.test(grid$arson,weights)
```

We have a moderate degree of spatial autocorrelation present in our data. This suggests that we need to use a modeling strategy that allows us to account for the impact of spatial autocorrelation.

## Modeling
### Poisson Model 1 - Main Effect

```{r}
model.poisson <- glm(arson ~ scale(delta70) , family = poisson, data = grid)

summary(model.poisson)
```


### Poisson Model 2 - Crime

```{r}
model.poisson2 <- glm(arson ~ scale(violent) + scale(property), family = poisson, data = grid)

summary(model.poisson2)
```

### Poisson Model 3 - Combined

```{r}
model.poisson3 <- glm(arson ~ scale(delta70) + scale(violent) + scale(property), family = poisson, data = grid)

summary(model.poisson3)
```

### Poisson Model 4 - 

```{r}
model.poisson4 <- glm(arson ~ scale(delta70) + scale(violent) + scale(property) + scale(prop_vacant) + 
                         scale(median_inc) + scale(black_prop) +
                        scale(pvty_prop) + scale(unemply_prop) + scale(ownOcc_prop), 
                      family = poisson, data = grid)

summary(model.poisson4)
```

### Check Dispersion

```{r}
dispersiontest(model.poisson4, alternative = "two.sided")
```

### Fit a Negative Binomial Model

```{r}
model.nb <- glm.nb(arson ~ scale(delta70) + + scale(prop_vacant) + scale(violent) + scale(property) + 
                         scale(black_prop) + scale(pvty_prop) + scale(unemply_prop) + scale(ownOcc_prop),
                   data = grid)
```



### Fit a Zero-Inflated Model
This uses a negative binomial distribution to compenstate for over dispersion. The zero-inflated aspect of the model is designed to model data with a large number of zeros, which is appropriate for these arson data because most grids have a count of 0.

```{r}
model.zoin1 <- zeroinfl(arson ~ scale(delta50) + scale(prop_vacant), 
                      dist = "negbin", data = grid)

summary(model.zoin1)
```

```{r}
model.zoin2 <- zeroinfl(arson ~ scale(violent) + scale(property), 
                      dist = "negbin", data = grid)

summary(model.zoin2)
```

```{r}
model.zoin3 <- zeroinfl(arson ~ scale(delta50) + scale(prop_vacant) + scale(violent) + scale(property), 
                      dist = "negbin", data = grid)

summary(model.zoin3)
```



```{r}
model.zoin4 <- zeroinfl(arson ~ scale(delta50) + scale(prop_vacant) + scale(violent) + scale(property) + 
                         scale(black_prop) + scale(pvty_prop) + scale(unemply_prop) + scale(ownOcc_prop), 
                      dist = "negbin", data = grid)

summary(model.zoin4)
```

```{r}
vuong(model.zoin4, model.nb)
```


Next, we'll fig a series of eigenvectors 

```{r}
me.fit <- spatialreg::ME(arson ~ scale(delta50) + scale(violent) + scale(property) + scale(prop_vacant) + 
                         scale(black_prop) + scale(pvty_prop) + scale(unemply_prop) + scale(ownOcc_prop), 
                      data = grid, family = negative.binomial(model.zoin4$theta), 
           listw = weights, verbose = TRUE, alpha = .05)
```

```{r}
me.fit
```

```{r}
model.zoin5 <- zeroinfl(arson ~ scale(delta50) + scale(prop_vacant) + scale(violent) + scale(property) + 
                         scale(black_prop) + scale(pvty_prop) + scale(unemply_prop) + scale(ownOcc_prop) +
                         fitted(me.fit), 
                      dist = "negbin", data = grid)

summary(model.zoin5)
```



```{r}
library(ggplot2)

ggplot(data = grid, mapping = aes(arson)) +
  geom_histogram(bins = 26)
```
