==> docs/99-analysis-spatial.Rmd <==
---
title: "Assignment Name"
author: "Your Name"
date: '(`r format(Sys.time(), "%B %d, %Y")`)'
output: 
  github_document: default
  html_notebook: default 
---

## Introduction
This notebook fits some experimental poisson and zero-inflated models.

## Dependencies
This notebook requires the following packages:

```{r load-packages}
# tidyverse packages
library(dplyr)

# spatial packages
library(sf)
library(sp)
library(spdep)

# modeling packages
library(MASS)
library(pscl)
library(INLA)

# other packages
library(here)
```

## Load Data
This notebook requires the final data set:

```{r load-data}
x <- st_read(here("data", "analysis", "spatial_analysis.geojson"), crs = 26915)
```

## Caluclate Pop Change

```{r}
x <- mutate(x, delta50 = (total_pop-pop50)/pop50*100)
x <- mutate(x, delta70 = total_pop-pop70/pop70*100)
```


## Weights Matrix

```{r}
weights <- poly2nb(x, queen=FALSE)
weights <- nb2listw(weights, style="W", zero.policy=TRUE)
```

Convert the spatial weight matrix into an INLA usable format:

```{r}
weights_alt <- poly2nb(x, queen=FALSE)
weights_alt <- nb2mat(weights_alt, style = "W") #W for row standarization
```

Create index in the one km dataset to match the spatial weight

```{r}
x$idx <- 1:nrow(x)
```

## Moran's I

```{r}
moran.test(x$arson,weights)
```


## Standard Poisson

```{r}
model.poisson <- glm(arson ~ violent + property, family = poisson, data = x)

summary(model.poisson)
```


```{r}
model.poisson2 <- glm(arson ~ violent + property + scale(prop_vacant) + 
                        scale(delta50) + median_inc + scale(black_prop) +
                        scale(pvty_prop) + scale(unemply_prop) + scale(ownOcc_prop), 
                      family = poisson, data = x)

summary(model.poisson2)
```

```{r}
model.poisson3 <- glm(arson ~ violent + property + (prop_vacant) + delta70, 
                      family = poisson, data = x)

summary(model.poisson3)
```

```{r}
lm.morantest(model.poisson3, listw = weights)
```

```{r}
AER::dispersiontest(model.poisson3, alternative = "two.sided")
```

## Negative Binomial

```{r}
model.nb <- glm.nb(arson ~ violent + property + (prop_vacant) + delta70, 
                      data = x)

summary(model.nb)
```

```{r}
round(exp(coef(model.nb)), 3)
```

```{r}
me.fit<-ME(arson ~ violent + property + (prop_vacant) + delta70, 
                      data = x, family = negative.binomial(1.729), listw = weights, verbose = TRUE, alpha = .05)
```

```{r}
me.fit
```


```{r}
model.nb2 <- glm.nb(arson ~ violent + property + (prop_vacant) + delta70 + fitted(me.fit), 
                      data = x)

summary(model.nb2)
```

```{r}
lm.morantest(model.nb, listw = weights)
lm.morantest(model.nb2, listw = weights)
```

```{r}
round(exp(coef(model.nb2)), 3)
```

## Zero-Inflated Poisson

```{r}
model.zoin <- zeroinfl(arson ~ violent + property + prop_vacant + black_prop, 
                      dist = "poisson", data = x)

summary(model.zoin)
```

## R-INLA

### Poisson

```{r}
formula <- arson ~ violent + property + prop_vacant + black_prop + f(idx, model="besag", graph = weights_alt)

poi_inla <- inla(formula, family = "poisson", data = x,
  control.compute = list(dic = TRUE),
  control.inla = list(tolerance = 1e-20, h = 1e-08), 
  control.predictor = list(compute = TRUE))

summary(poi_inla)
```

### Zero Inflated

```{r}
zero_inla <- inla(formula, family = "zeroinflatedpoisson0", data = x,
  control.compute = list(dic = TRUE),
  control.inla = list(tolerance = 1e-20, h = 1e-08), 
  control.predictor = list(compute = TRUE))

summary(zero_inla)
```
